+++ R/mirai-package.R
@@ -49,7 +49,7 @@
 #'     (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})
 #'
 #' @importFrom nanonext base64dec call_aio context is_error_value mclock msleep
-#'     parse_url random recv recv_aio request send socket stat stop_aio
+#'     opt parse_url random recv recv_aio request send socket stat stop_aio
 #'     unresolved
 #'
 #' @docType package